Split Subaru legacy to separate file
